+++ src/extension.ts
@@ -5,6 +5,7 @@ import * as dds from './other/dds';
 import * as rdp from './other/rdp';
 import { AssetsSymbolProvider } from './features/outline/assetsSymbolProvider';
 import { AssetsActionProvider } from './features/assetsActionProvider';
+import * as AssetsDecorator from './features/assetsDecorator';
 import * as commands from './features/commands';
 
 import * as logger from './other/logger';
@@ -25,6 +26,7 @@ export function activate(context: vscode.ExtensionContext) {
 	context.subscriptions.push(...AssetsActionProvider.register(context));
 	context.subscriptions.push(...registerGuidUtilsProvider(context));
 	context.subscriptions.push(...cfgHoverProvider.registerHoverProvider(context));
+	AssetsDecorator.activate(context);
 
 	commands.registerCommands(context);
 }